move isServicePrincipalSetInEnv() into azureUtils
